requirements.txt
absl-py==0.11.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1613038019788/work
astunparse==1.6.3
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
basemap==1.2.0
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
branca==0.4.2
brotlipy==0.7.0
cachetools==4.2.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247308275/work
chardet==4.0.0
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography @ file:///C:/ci/cryptography_1616769432139/work
cycler==0.10.0
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dill==0.3.3
entrypoints==0.3
flatbuffers==1.12
folium==0.12.1
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1609784788857/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1615900693889/work
importlib-resources==5.1.2
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1614616717881/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1610481889018/work
jedi @ file:///C:/ci/jedi_1606932641827/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
joblib==1.0.1
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213516947/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1612282618948/work
Markdown==3.3.3
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1601914921407/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///C:/ci/notebook_1616443616158/work
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas @ file:///C:/ci/pandas_1614711338358/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso==0.7.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1615224334542/work
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyproj==1.9.6
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
pyshp @ file:///tmp/build/80754af9/pyshp_1610641829514/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole @ file:///tmp/build/80754af9/qtconsole_1616775094278/work
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.24.1
scipy @ file:///C:/ci/scipy_1616703425152/work
seaborn @ file:///tmp/build/80754af9/seaborn_1608578541026/work
selenium @ file:///C:/ci/selenium_1594408106746/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
six @ file:///C:/ci/six_1605205426665/work
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu==2.4.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-metadata==0.29.0
termcolor==1.1.0
terminado==0.9.3
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
threadpoolctl==2.1.0
tornado @ file:///C:/ci/tornado_1606935947090/work
tqdm==4.59.0
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
